Adding support for Arrays (#151)
* initial WIP for adding full Array support

* update columns to use array a bit cleaner. Comment out other array fields for now until I get Array(Bool) working. After that, the rest should be easy

* fixing syntax error and typo

* decided to work with Array(String) instead. getting closer

* formatting

* this is a hack, but specs are passing now

* re-enable the other array fields. Cleaned up some weird confusion with charms and parse methods.

* moved Array(String) parsing up a level. specs are passing

* added in support for alter table. Re-enabled the BucketBox data generation. Added some array parsing specs

* removing initial test support for Array(Time) and Array(UUID) for now until the pg shard has support for them

* working on saving arrays in to the DB. This commit still contains a compilation error

* removing Array(Float64) support for now since there's an issue with the pg shard. Also fixed #148

* moves the hack down a level which simplifies some logic

* all Float fields must be Float64 or Float32 for now. Crystal has some issues when it's more generic like Float. Also made migrator columns Generic now which helps enforce the type.

* can never go wrong with a few more specs

* float is now abstract

+++ db/migrations/20180113081408_create_companies.cr
@@ -4,7 +4,7 @@ class CreateCompanies::V20180113081408 < Avram::Migrator::Migration::V1
       primary_key id : Int64
       add_timestamps
       add sales : Int64
-      add earnings : Float
+      add earnings : Float64
     end
   end
